[#4] Iterate on review feedback
