+++ src/index.less
@@ -1,19 +1,17 @@
-// ## Deps
+// Dependencies
 @import "font-awesome";
 @import "bootstrap";
 @import "octicons";
 
+// Configuration
 @import "./less/variables.less";
-@import "./less/xl.less";
 
-/**
- * ## Components
- */
-@import "./less/dropdown-menu.less";
-@import "./less/page-header.less";
+// Tweaks to bootstrap
+@import "./less/xl.less";
 @import "./less/panel.less";
-@import "./less/message-background.less";
-@import "./less/sidebar.less";
+
+// Components
+@import "./src/repos/index.less";
 
 html, body{
   height: 100%;
@@ -29,9 +27,7 @@ html, body{
 }
 
 .content {
-  // .container-fluid;
   margin-left: 220px;
-  // margin-right: 220px;
   height: 100%;
   &.no-sidebar {
     margin-left: 0;
@@ -49,63 +45,5 @@ html, body{
 
   p {
     text-align: center;
-  }
-}
-
-
-.grid-item {
-  // min-width: 259px;
-  max-height: 167px;
-  overflow: hidden;
-  margin-bottom: 20px;
-  a {
-    .thumbnail;
-    display: block;
-    padding-bottom: 0;
-    border-bottom-right-radius: 0;
-    border-bottom-left-radius: 0;
-    .grid-item-heading {
-      .h5;
-      margin-top: 0;
-      font-weight: bold;
-      padding-top: 10px;
-      padding-left: 3%;
-      white-space: nowrap;
-    }
-    .avatar-container{
-      height: 128px;
-      width: 128px;
-      overflow-y: hidden;
-      margin-left: 7%;
-      margin-right: auto;
-    }
-    img.avatar {
-      margin-top: 30px;
-      position: absolute;
-      -webkit-transition: all 500ms cubic-bezier(0.680, 0, 0.265, 1); /* older webkit */
-      -webkit-transition: all 500ms cubic-bezier(0.680, -0.550, 0.265, 1.550);
-         -moz-transition: all 500ms cubic-bezier(0.680, -0.550, 0.265, 1.550);
-           -o-transition: all 500ms cubic-bezier(0.680, -0.550, 0.265, 1.550);
-              transition: all 500ms cubic-bezier(0.680, -0.550, 0.265, 1.550); /* easeInOutBack */
-
-      -webkit-transition-timing-function: cubic-bezier(0.680, 0, 0.265, 1); /* older webkit */
-      -webkit-transition-timing-function: cubic-bezier(0.680, -0.550, 0.265, 1.550);
-         -moz-transition-timing-function: cubic-bezier(0.680, -0.550, 0.265, 1.550);
-           -o-transition-timing-function: cubic-bezier(0.680, -0.550, 0.265, 1.550);
-              transition-timing-function: cubic-bezier(0.680, -0.550, 0.265, 1.550); /* easeInOutBack */
-    }
-
-    img[data-hook='travis_badge_url'] {
-      .pull-right;
-      margin-top: 8px;
-    }
-
-
-    &:hover {
-      text-decoration: none;
-      img.avatar {
-        margin-top: 5px;
-      }
-    }
   }
 }
