--- src/less/dropdown-menu.less
@@ -1,16 +1,0 @@
-ul.dropdown-menu {
-  max-height: 200px;
-  overflow-y: auto;
-}
-
-.dropdown-menu-label {
-  .text-muted;
-  margin-right: 10px;
-  display: inline-block;
-}
-
-.dropdown-menu-active {
-  font-weight: bold;
-  display: inline-block;
-  margin-right: 10px;
-}
