--- src/less/message-background.less
@@ -1,32 +1,0 @@
-ul.message-background {
-  font-weight: bold;
-  color: rgba(0, 0, 0, 0.2);
-  font-size: (@font-size-base * 3);
-  margin: 0;
-  padding: 0;
-}
-
-ul.message-background.centered {
-  display: -webkit-flex;
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  -webkit-align-items: center;
-  text-align: center;
-}
-
-ul.message-background.centered li {
-  width: 100%;
-}
-ul.message-background li {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-}
-
-.message-background-tip {
-  font-size: @font-size-base;
-  margin-top: (@font-size-base * 3);
-}
