--- src/less/page-header.less
@@ -1,21 +1,0 @@
-.page-header {
-  // font-size: 24px;
-  // margin: 0 0 20px;
-  // padding: 0;
-  // border: none;
-
-  i.fa {
-    margin-right: 5px;
-  }
-  i.fa + span {
-    margin-right: 5px;
-  }
-
-  .btn-toolbar {
-    padding-top: 5px;
-  }
-
-  .pull-right > .btn-toolbar > .btn-group:last-child > ul.dropdown-menu {
-    margin-left: -15px;
-  }
-}
