+++ src/less/panel.less
@@ -1,5 +1,4 @@
 @import "./variables.less";
-@import "./type.less";
 
 .panel {
   .list-group {
@@ -9,18 +8,6 @@
       border-width: 1px 0;
       border-radius: 0;
     }
-    // &:first-child {
-    //   .list-group-item:first-child {
-    //     border-top: 0;
-    //     .border-top-radius((@panel-border-radius - 1));
-    //   }
-    // }
-    // &:last-child {
-    //   .list-group-item:last-child {
-    //     border-bottom: 0;
-    //     .border-bottom-radius((@panel-border-radius - 1));
-    //   }
-    // }
   }
   .btn-group-justified {
     a.btn {
