+++ src/less/variables.less
@@ -6,75 +6,14 @@
 //== Colors
 //
 //## Gray and brand colors for use across Bootstrap.
-@green: #00AC51;
-@red: #ff5b57;
-@blue: #428bca;
-@darker-black: #242a30;
-@black: #2d353c;
-@white: #FFFFFF;
-
-@white: #D9D9D9;
-@very-light-gray: #80807D;
-@light-gray: #75715E;
-@gray: #3E3D32;
-@dark-gray: #3A3B36;
-@very-dark-gray: #272822;
-
-@cyan: #6CC7D9;
-@blue: #6CC7D9;
-@purple: #A082D9;
-@pink: #E52D71;
-@orange: #D9882A;
-@light-orange: #CCC47A;
-
-@background-positive: @green;
-@background-negative: @red;
-@background-neutral: @gray-dark;
-
-@syntax-text-color: @white;
-@syntax-cursor-color: #F1F2F3;
-@syntax-selection-color: @gray;
-@syntax-background-color: @very-dark-gray;
-
-// Guide colors
-@syntax-wrap-guide-color: @dark-gray;
-@syntax-indent-guide-color: @dark-gray;
-@syntax-invisible-character-color: @dark-gray;
-
-// For find and replace markers
-@syntax-result-marker-color: #ACACA9;
-@syntax-result-marker-color-selected: #FFE792;
-
-// Gutter colors
-@syntax-gutter-text-color: @very-light-gray;
-@syntax-gutter-text-color-selected: @syntax-gutter-text-color;
-@syntax-gutter-background-color: @syntax-background-color;
-@syntax-gutter-background-color-selected: @syntax-selection-color;
-
-// For git diff info. i.e. in the gutter
-@syntax-color-renamed: @blue;
-@syntax-color-added: @green;
-@syntax-color-modified: @orange;
-@syntax-color-removed: @red;
-
 @gray-darker:            lighten(#000, 13.5%); // #222
 @gray-dark:              lighten(#000, 20%);   // #333
 @gray:                   lighten(#000, 33.5%); // #555
 @gray-light:             lighten(#000, 46.7%); // #777
 @gray-lighter:           lighten(#000, 93.5%); // #eee
 
-// @brand-primary:         @green;
-// @brand-success:         @green;
-// @brand-info:            #5bc0de;
-// @brand-warning:         #f0ad4e;
-// @brand-danger:          @red;
-
-// @background-positive: @green;
-// @background-negative: @red;
-// @background-neutral: @gray-dark;
-
 // //== Scaffolding
-@body-bg:              #f8f9fd;
+@body-bg:               #f8f9fd;
 @text-color:            #677587;
 @link-color:            #677587;
 @link-hover-color:      darken(@link-color, 15%);
