+++ src/repos/grid-item.less
@@ -1,0 +1,55 @@
+.grid-item {
+  max-height: 167px;
+  overflow: hidden;
+  margin-bottom: 20px;
+  a {
+    .thumbnail;
+    display: block;
+    padding-bottom: 0;
+    border-bottom-right-radius: 0;
+    border-bottom-left-radius: 0;
+    .grid-item-heading {
+      .h5;
+      margin-top: 0;
+      font-weight: bold;
+      padding-top: 10px;
+      padding-left: 3%;
+      white-space: nowrap;
+    }
+    .avatar-container{
+      height: 128px;
+      width: 128px;
+      overflow-y: hidden;
+      margin-left: 7%;
+      margin-right: auto;
+    }
+    img.avatar {
+      margin-top: 30px;
+      position: absolute;
+      -webkit-transition: all 500ms cubic-bezier(0.680, 0, 0.265, 1); /* older webkit */
+      -webkit-transition: all 500ms cubic-bezier(0.680, -0.550, 0.265, 1.550);
+         -moz-transition: all 500ms cubic-bezier(0.680, -0.550, 0.265, 1.550);
+           -o-transition: all 500ms cubic-bezier(0.680, -0.550, 0.265, 1.550);
+              transition: all 500ms cubic-bezier(0.680, -0.550, 0.265, 1.550); /* easeInOutBack */
+
+      -webkit-transition-timing-function: cubic-bezier(0.680, 0, 0.265, 1); /* older webkit */
+      -webkit-transition-timing-function: cubic-bezier(0.680, -0.550, 0.265, 1.550);
+         -moz-transition-timing-function: cubic-bezier(0.680, -0.550, 0.265, 1.550);
+           -o-transition-timing-function: cubic-bezier(0.680, -0.550, 0.265, 1.550);
+              transition-timing-function: cubic-bezier(0.680, -0.550, 0.265, 1.550); /* easeInOutBack */
+    }
+
+    img[data-hook='travis_badge_url'] {
+      .pull-right;
+      margin-top: 8px;
+    }
+
+
+    &:hover {
+      text-decoration: none;
+      img.avatar {
+        margin-top: 5px;
+      }
+    }
+  }
+}
