+++ src/repos/grid.js
@@ -1,8 +1,8 @@
 var AmpersandView = require('ampersand-view'),
-  RepoGridItemView = require('./repo-grid-item.js');
+  RepoGridItemView = require('./grid-item.js');
 
 module.exports = AmpersandView.extend({
-  template: require('./repo-grid.jade'),
+  template: '<div data-hook="repo-grid"></div>',
   render: function() {
     this.renderWithTemplate();
     this.renderCollection(this.collection, RepoGridItemView, this.queryByHook('repo-grid'));
