+++ src/repos/index.less
@@ -1,3 +1,5 @@
+@import "./grid-item.less";
+
 .sidebar {
   position: absolute;
   z-index: 10;
@@ -9,11 +11,8 @@
   }
   .facet-name {
     .h5;
-    // font-weight: bold;
     font-weight: 600;
     padding-left: 10px;
-    // border-top: 1px solid @well-border;
-    // padding-top: 10px;
   }
 
   .facet-value-list{
@@ -50,20 +49,8 @@
   }
 }
 
-.sidebar-collapsed {
-  // .sidebar-bg {
-  //   width: 24px;
-  //   overflow: hidden;
-  //   display: none;
-  // }
-  // .sidebar {
-  //   display: none;
-  // }
-}
-
 .sidebar label {
   font-weight: 600;
-  // font-size: 90%;
 }
 
 
@@ -90,9 +77,7 @@
 }
 
 .sidebar, .sidebar-bg {
-  // top: 80px;
   top: 0px;
-  // right: 0;
   left: 0;
   bottom: 0;
   width: 220px;
