+++ src/repos/list-item.jade
@@ -16,14 +16,4 @@ a(data-hook='url', target='_blank').list-group-item
         li(style='width: 90px;')
           img(data-hook='travis_badge_url')
 
-        li
-          span(data-hook='language')
-        li
-          span(data-hook='category')
-        li
-          span(data-hook='size')
-        li
-          span(data-hook='last_activity_at_string')
-        li
-          span(data-hook='created_at_string')
       p(data-hook='description')
