+++ src/repos/list-item.js
@@ -1,0 +1,5 @@
+var ItemView = require('./item');
+
+module.exports = ItemView.extend({
+  template: require('./list-item.jade')
+});
