+++ src/repos/list.js
@@ -1,8 +1,8 @@
 var AmpersandView = require('ampersand-view'),
-  RepoListItemView = require('./repo-list-item.js');
+  RepoListItemView = require('./list-item.js');
 
 module.exports = AmpersandView.extend({
-  template: require('./repo-list.jade'),
+  template: '<div class="list-group" data-hook="repo-list"></div>',
   render: function() {
     this.renderWithTemplate();
     this.renderCollection(this.collection, RepoListItemView, this.queryByHook('repo-list'));
