--- src/views/repo-grid.jade
@@ -1,1 +1,0 @@
-div(data-hook='repo-grid')
