--- src/views/repo-list.jade
@@ -1,1 +1,0 @@
-.list-group(data-hook='repo-list')
